Add PostSetUserHook error-handling
In the event that the PostSetUserHook throws an error, we should treat
the set_user call as erroneous and properly back out. To do this, move
the reset_user code out to a separate function and wrap the
PostSetUserHook in a PG_TRY(); block, so we can perform a reset and then
rethrow the error.